asdf

+++ src/core/tensor.py
@@ -2,7 +2,7 @@ from src.utils.backend import xp
 
 class Tensor:
     def __init__(self, data, requires_grad=True, requires_mask=False, name=None, eps=1e-4):
-        self.data = xp.array(data).astype(xp.float64)
+        self.data = xp.array(data).astype(xp.float16)
         self.requires_grad = requires_grad
         self.parents = ()
         self.grad_fn = None
